+++ udata/core/dataset/models.py
@@ -14,12 +14,13 @@ from mongoengine.signals import post_save, pre_save
 from stringdist import rdlevenshtein
 from werkzeug.utils import cached_property
 
+from udata.api_fields import field
 from udata.app import cache
 from udata.core import storages
 from udata.core.owned import Owned, OwnedQuerySet
 from udata.frontend.markdown import mdstrip
 from udata.i18n import lazy_gettext as _
-from udata.models import BadgeMixin, SpatialCoverage, WithMetrics, db
+from udata.models import Badge, BadgeMixin, BadgesList, SpatialCoverage, WithMetrics, db
 from udata.mongo.errors import FieldValidationError
 from udata.uris import ValidationError, endpoint_for
 from udata.uris import validate as validate_url
@@ -52,6 +53,10 @@ __all__ = (
     "CommunityResource",
     "ResourceSchema",
 )
+
+BADGES: dict[str, str] = {
+    PIVOTAL_DATA: _("Pivotal data"),
+}
 
 NON_ASSIGNABLE_SCHEMA_TYPES = ["datapackage"]
 
@@ -498,7 +503,16 @@ class Resource(ResourceMixin, WithMetrics, db.EmbeddedDocument):
         self.dataset.save(*args, **kwargs)
 
 
-class Dataset(WithMetrics, BadgeMixin, Owned, db.Document):
+class DatasetBadge(Badge):
+    kind = db.StringField(required=True, choices=list(BADGES.keys()))
+
+
+class DatasetBadgeMixin(BadgeMixin):
+    badges = field(BadgesList(DatasetBadge), **BadgeMixin.default_badges_list_params)
+    __badges__ = BADGES
+
+
+class Dataset(WithMetrics, DatasetBadgeMixin, Owned, db.Document):
     title = db.StringField(required=True)
     acronym = db.StringField(max_length=128)
     # /!\ do not set directly the slug when creating or updating a dataset
@@ -538,10 +552,6 @@ class Dataset(WithMetrics, BadgeMixin, Owned, db.Document):
 
     def __str__(self):
         return self.title or ""
-
-    __badges__ = {
-        PIVOTAL_DATA: _("Pivotal data"),
-    }
 
     __metrics_keys__ = [
         "discussions",